requirements.txt
Django
djangorestframework
requests
django-redis